Allow builds that depend on OpenGL.
If the environment variable 'WITH_OPENGL' is set, cmake will try to find
OpenGL and GLUT libraries and development headers. In case this is
successful, the CPP flag 'WITH_OPENGL' will be defined and the build is
set up to use OpenGL. Otherwise, a warning message that OpenGL is not
supported will be issued.