Do not use glFinish() but glFlush().
Using glFinish() after a buffer swap causes a significant CPU overhead.
For instance, the title screen used 30% of a CPU core for rendering with
glFinish(). With glFlush() it merely use 3% of a core. Since the main
loop uses SDL ticks to syncronize glFinish() should not be needed
anyway.

This commit also includes a printout of the OpenGL renderer and driver
information on construnction of the OpenGLRenderDevice.