Introduce local frame for renderables.
Renderables now have a field local_frame of type SDL_Rect. This allows
to define their target coordinates in a local frame and the render
device will take care of the transformation to global. If the local
frame defines a width and heigth in addition to the origin, correct
clipping will be performed by the render device. This enables widget
child rendering to be done without wasteful pre-rendering to an
SDL_Surface and benefits all render devices. It potentially also allows
front end code to be simplified in many other places (like menus).