Never render widgets to SDL_Surfaces. Fixes #1.
The widget render interface has been altered to not accept an
SLD_Surface as a target. Instead, widgets are now aware of their local
frame in case they are child widgets. The proper transformation to
screen coordinates and clipping is handle by the render device.

This commit also contains some performance improvement. WidgetLabels are
no longer created on the fly in other widgets. Instead, they are all
cached and updates to their content is only triggered when necessary.