Implement texture cache for animations.
It turns out the ownership of animations is a complicated issue. The
current implementation is rather wasteful, creating deep copies all the
time, while the only difference between entities sharing an animation is
the current frame and such. No need to copy all resources around.
Another re-factoring session on the horizon there. For now, stick with
the current paradigm, but give ownership of graphics resources to the
animation set.

Good news: with this change the game state renders even more efficiently in
OpenGL. Yay.

This fixes #2.